% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_lists.R
\name{compare_old_new_library}
\alias{compare_old_new_library}
\title{compare lists}
\usage{
compare_old_new_library()
}
\value{
matrix
}
\description{
Compares the old and new package lists, and then returns the list of things
that have yet to be installed in the new library location.
}
