% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_and_get_options.R
\name{get_new_library}
\alias{get_new_library}
\title{get new library location}
\usage{
get_new_library()
}
\value{
character
}
\description{
returns the new library location that was defined by \code{set_new_library()}
}
