% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_lists.R
\name{get_new_packages}
\alias{get_new_packages}
\title{new packages}
\usage{
get_new_packages(new_library = get_new_library())
}
\arguments{
\item{new_library}{where to look}
}
\value{
matrix
}
\description{
get the list of packages installed in the new library
}

