% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_and_get_options.R
\name{get_old_library}
\alias{get_old_library}
\title{get the old library location}
\usage{
get_old_library()
}
\value{
character
}
\description{
Gets the old library location set by \code{set_old_library()}.
}
