% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_lists.R
\name{get_old_packages}
\alias{get_old_packages}
\title{old packages}
\usage{
get_old_packages(old_library = get_old_library())
}
\arguments{
\item{old_library}{where to look}
}
\value{
matrix
}
\description{
get the list of packages installed in the old library
}

