% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reup.R
\name{install_packages}
\alias{install_packages}
\title{install packages}
\usage{
install_packages(pkg_frame, install_fun)
}
\arguments{
\item{pkg_frame}{data.frame of packages to be installed}

\item{install_fun}{the installation function to use}
}
\value{
data.frame
}
\description{
Try to install a list of packages in a \code{data.frame}. Installation success
is determined by comparing to results of \code{installed.packages}, so the list
of uninstalled may not be exactly correct if a package was installed previously.
}
