% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_lists.R
\name{n_package_deps}
\alias{n_package_deps}
\title{number of dependencies}
\usage{
n_package_deps(pkg_matrix)
}
\arguments{
\item{pkg_matrix}{the matrix of package information}
}
\value{
data.frame
}
\description{
get the number of dependencies for all of the packages
}
