% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_lists.R
\name{package_type}
\alias{package_type}
\title{package type}
\usage{
package_type(pkg_frame)
}
\arguments{
\item{pkg_frame}{a data.frame of package information}
}
\value{
data.frame
}
\description{
Determine what type of package it is and where it needs to be installed from.
The types expected include: Cran, Bioconductor, local, and versioned repo (i.e.
github, bitbucket, gitlab).
}

