% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_lists.R
\docType{data}
\name{pkg_matrix}
\alias{pkg_matrix}
\title{pkg matrix}
\format{An object of class \code{matrix} with 283 rows and 16 columns.}
\source{
Robert M Flight
}
\usage{
pkg_matrix
}
\description{
An example package information matrix generated by RMF on his personal R
installation.
}
\keyword{datasets}
