% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reup.R
\name{reup}
\alias{reup}
\title{reup}
\usage{
reup(...)
}
\arguments{
\item{...}{various options, normally set using the setter functions}
}
\description{
actually tries to do the upgrade of packages using `install.packages`.
}
