% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_and_get_options.R
\name{set_bioc_mirror}
\alias{set_bioc_mirror}
\title{set bioconductor mirror}
\usage{
set_bioc_mirror(url = "https://bioconductor.org")
}
\description{
Sets the Bioconductor mirror to use. Default is https://bioconductor.org
}

