% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reup.R
\name{set_cran_mirror}
\alias{set_cran_mirror}
\title{set cran mirror}
\usage{
set_cran_mirror(url = "https://cloud.r-project.org/")
}
\arguments{
\item{url}{the url of the CRAN mirror}
}
\description{
Sets the CRAN mirror to use. Default is the RStudio one
}

