% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_and_get_options.R
\name{set_local_packages}
\alias{set_local_packages}
\title{set local packages}
\usage{
set_local_packages(local_pkg_dir = getwd())
}
\arguments{
\item{local_pkg_dir}{the directory to search for local packages}
}
\description{
Set a directory to search recursively for the existence of local packages
that should be installed using \code{devtools}.
}

