% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reup.R
\name{set_new_library}
\alias{set_new_library}
\title{new library location}
\usage{
set_new_library(parent_directory = NULL, library_directory = NULL)
}
\arguments{
\item{parent_directory}{the parent directory}

\item{library_directory}{the actual library directory, default is NULL}
}
\description{
Creates a new library location based on the version of R, and if the
Bioconductor mirror is set, then the Bioconductor version as well.
}

