% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_and_get_options.R
\name{set_old_library}
\alias{set_old_library}
\title{old library location}
\usage{
set_old_library(lib_loc = NULL)
}
\arguments{
\item{lib_loc}{the library location}
}
\description{
Sets the old library. Default is to take it from R_LIBS
}
