% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reup.R
\name{setup_devtools}
\alias{setup_devtools}
\title{setup devtools}
\usage{
setup_devtools()
}
\description{
Makes modifications suitable for doing installations using \code{devtools} for
local and / or github based packages.
}

