% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_and_get_options.R
\name{set_new_library}
\alias{set_new_library}
\title{new library location}
\usage{
set_new_library(lib_loc = NULL)
}
\arguments{
\item{lib_loc}{the full directory name}
}
\description{
Creates a new library location based on the version of R, and if the
Bioconductor mirror is set, then the Bioconductor version as well.
}

